requirements.txt
# MatrixPilot 生产依赖

# Flask核心
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Caching>=2.0.0

# API文档
flask-restx>=1.3.0

# 数据库
SQLAlchemy>=2.0.0

# 定时任务
APScheduler>=3.10.0

# HTTP请求
requests>=2.28.0

# 环境变量
python-dotenv>=1.0.0

# 生产服务器
gunicorn>=21.0.0
